% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_ints}
\alias{clean_ints}
\title{Removes sign information from interactions}
\usage{
clean_ints(x)
}
\arguments{
\item{x}{A character vector of interactions}
}
\value{
A character vector of interactions without sign information
}
\description{
Removes sign information from interactions
}
\keyword{internal}
