% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stability.R
\name{get_top_stable_features}
\alias{get_top_stable_features}
\title{Get most stable features according to average local stability importance}
\usage{
get_top_stable_features(
  rf_fit,
  X,
  y,
  k = NULL,
  min_stability = NULL,
  features = NULL,
  feature_groups = NULL,
  first_only = FALSE
)
}
\arguments{
\item{rf_fit}{Fitted random forest object. Must be of class "ranger" (i.e.,
the output of \code{\link[ranger:ranger]{ranger::ranger()}}).}

\item{X}{Data frame of sample design matrix, used to evaluate local feature
stability.}

\item{y}{Vector of observed responses in the same order as the rows of
\code{X}. Must be binary vector with 0s and 1s.}

\item{k}{Number of top features to return. Must provide either \code{k} or
\code{min_stability}. If both \code{k} and \code{min_stability} are
\code{NULL}, a data frame with all features along with their average local
stability importance is returned.}

\item{min_stability}{Minimum threshold for the average local stability
importance. Any feature with a lower average local stability importance
will be filtered out. Only used if \code{k} is \code{NULL}. Must provide
either \code{k} or \code{min_stability}. If both \code{k} and
\code{min_stability} are \code{NULL}, a data frame with all features along
with their average local stability importance is returned.}

\item{features}{Vector of (signed) features, for which to evaluate stability.
Default (\code{NULL}) is to evaluate the local feature stability for all
features in \code{X}.}

\item{feature_groups}{Data frame of feature-to-group/superfeature mapping.
Should have the two columns: "feature" and "group". Default (\code{NULL})
performs no grouping so that each feature is its own group.}

\item{first_only}{Logical; whether or not to include only the first
appearance of the feature in a tree when computing local stability scores.
Default (\code{FALSE}) includes all occurrences of a feature in the
local stability score computation.}
}
\value{
If \code{k} and \code{min_stability} are \code{NULL}, returns data
frame with all features along with their average local stability
importance, which can be used to rank the features manually. Otherwise,
returns vector of the most stable features.
}
\description{
Get most stable features according to average local stability importance
}
