% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stability.R
\name{local_rf_feature_stability}
\alias{local_rf_feature_stability}
\title{Compute local feature stability scores in a random forest.}
\usage{
local_rf_feature_stability(
  rf_fit,
  X,
  features = NULL,
  feature_groups = NULL,
  first_only = FALSE
)
}
\arguments{
\item{rf_fit}{Fitted random forest object. Must be of class "ranger" (i.e.,
the output of \code{\link[ranger:ranger]{ranger::ranger()}}).}

\item{X}{Data frame of sample design matrix, used to evaluate local feature
stability.}

\item{features}{Vector of features, for which to evaluate stability.
Default (\code{NULL}) is to evaluate the local feature stability for all
features in \code{X}.}

\item{feature_groups}{Data frame of feature-to-group/superfeature mapping.
Should have the two columns: "feature" and "group". Default (\code{NULL})
performs no grouping so that each feature is its own group.}

\item{first_only}{Logical; whether or not to include only the first
appearance of the feature in a tree when computing local stability scores.
Default (\code{FALSE}) includes all occurrences of a feature in the
local stability score computation.}
}
\value{
A data frame of the same size as \code{X} containing the local
feature stability scores from the fitted random forest.
}
\description{
Compute local feature stability scores in a random forest.
}
