% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stability.R
\name{local_rf_int_stability}
\alias{local_rf_int_stability}
\title{Compute local interacttion stability scores in a random forest.}
\usage{
local_rf_int_stability(
  rf_fit,
  X,
  ints,
  feature_groups = NULL,
  first_only = FALSE
)
}
\arguments{
\item{rf_fit}{Fitted random forest object. Must be of class "ranger" (i.e.,
the output of \code{\link[ranger:ranger]{ranger::ranger()}}).}

\item{X}{Data frame of sample design matrix, used to evaluate local feature
stability.}

\item{ints}{Vector of (signed) interactions to evaluate local stability.}

\item{feature_groups}{Data frame of feature-to-group/superfeature mapping.
Should have the two columns: "feature" and "group". Default (\code{NULL})
performs no grouping so that each feature is its own group.}

\item{first_only}{Logical; whether or not to include only the first
appearance of the feature in a tree when computing local stability scores.
Default (\code{FALSE}) includes all occurrences of a feature in the
local stability score computation.}
}
\value{
A data frame of the same size as \code{X} containing the local
interaction stability scores from the fitted random forest.
}
\description{
Compute local interacttion stability scores in a random forest.
}
