% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stability.R
\name{local_rf_stability_importance}
\alias{local_rf_stability_importance}
\title{Compute the local stability importance scores for features and interactions
in a random forest}
\usage{
local_rf_stability_importance(
  rf_fit,
  X,
  y,
  features = NULL,
  ints = NULL,
  feature_groups = NULL,
  first_only = FALSE,
  nperm = 10000
)
}
\arguments{
\item{rf_fit}{Fitted random forest object. Must be of class "ranger" (i.e.,
the output of \code{\link[ranger:ranger]{ranger::ranger()}}).}

\item{X}{Data frame of sample design matrix, used to evaluate local feature
stability.}

\item{y}{Vector of observed responses in the same order as the rows of
\code{X}. Must be binary vector with 0s and 1s.}

\item{features}{Vector of (signed) features, for which to evaluate stability.
Default (\code{NULL}) is to evaluate the local feature stability for all
features in \code{X}.}

\item{ints}{Vector of (signed) interactions to evaluate local stability.}

\item{feature_groups}{Data frame of feature-to-group/superfeature mapping.
Should have the two columns: "feature" and "group". Default (\code{NULL})
performs no grouping so that each feature is its own group.}

\item{first_only}{Logical; whether or not to include only the first
appearance of the feature in a tree when computing local stability scores.
Default (\code{FALSE}) includes all occurrences of a feature in the
local stability score computation.}

\item{nperm}{Number of permutations used in the permutation test.}
}
\value{
A list with the following components:
\describe{
\item{feature_stability_df}{Data frame of size (\code{nrow(X)}) x
(\code{length(features)}) containing the local feature stability scores
from the fitted random forest.}
\item{feature_stability_pvals}{Tibble of p-values from the permutation test
for each feature in \code{feature_stability_df}.}
\item{int_stability_df}{Data frame of size (\code{nrow(X)}) x
(\code{length(ints)}) containing the local interaction stability scores
from the fitted random forest.}
\item{int_stability_pvals}{Tibble of p-values from the permutation test
for each interaction in \code{int_stability_df}.}
}
}
\description{
Compute the local stability importance scores for features and interactions
in a random forest
}
