% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_signs}
\alias{remove_signs}
\title{Removes sign information from features/interactions}
\usage{
remove_signs(x)
}
\arguments{
\item{x}{A character vector of signed features or interactions}
}
\value{
A character vector of interactions without sign information
}
\description{
Removes sign information from features/interactions
}
\keyword{internal}
