% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_stability.R
\name{twosample_permutation_test}
\alias{twosample_permutation_test}
\title{Run a two-sample permutation test with the difference-in-means test statistic.}
\usage{
twosample_permutation_test(local_stability_df, y, feature, nperm = 10000)
}
\arguments{
\item{local_stability_df}{Data frame with the local stability scores.
Generally the output of \code{local_rf_feature_stability()} or
\code{local_rf_int_stability()}.}

\item{y}{Vector of observed responses in the same order as the rows of
\code{local_stability_df}. Must be binary vector with 0s and 1s.}

\item{feature}{Character string indicating the feature/interaction to test.}

\item{nperm}{Number of permutations used in the permutation test.}
}
\value{
A list of 3:
\describe{
\item{pval}{Two-sided permutation p-value.}
\item{T_obs}{Observed difference-in-means test statistic.}
\item{perm_dist}{Null permutation distribution.}
}
}
\description{
Run a two-sample permutation test with the difference-in-means test statistic.
}
